requirements.txt
google-cloud-aiplatform
kfp
google-cloud-pipeline-components
gradio
google-cloud-vision